BinaryKmer changed to struct;  Added inferred kmers to path_format; renamed binary_*
BinaryKmer is a struct instead of just a typedef array. This makes it possible
to pass by copy and to return a BinaryKmer. Also allows the removal of
BinaryKmerPtr and ConstBinaryKmerPtr types. It should now be clearer when a
variable is an array or a pointer.

Inferred kmers were added to path_format.c and add_read_paths.c. They are dumped
when adding paths and loaded when loading a ctp with a ctx file. Paths can now
be added using a smaller background population than will later be used.

graph_walker:
- fixed graph_walker so paths are only used for forks in the colour being followed

path_format.c:
- faster memory arrangement: linked lists arranged in forward direction before
  dumping. This should improve graph traversal speed.

hash_table.c:
- Changed how memory is decided for hash table so we are consistent
- Added WARN_OCCUPANCY for use when deciding memory

add_read_paths.c:
- changes to improve multi-threading
- changes to support writing inferred kmers

Renamed files / methods:
- renamed binary_paths -> paths_store
- renamed binary_format -> graph_format
- renamed binary_reader -> graph_reader
- renamed binary_writer -> graph_writer

cmd:
 * multiple [-p <in.ctp>] arguments accepted