Fix breakpoint caller bug when using only one colour
src/commands/ctx_breakpoints.c:
src/tools/kmer_occur.c:
src/tools/breakpoint_caller.c:
Add the reference as its own colour so that there are always >1 colours for
breakpoint calling. This avoids the scenario where we think we have only 1
colour and so do not bother to check that kmers are in our colour. This was
causing GraphWalker to abort. In future, all kmers should belong to a colour!

results/bubble_caller_10ecoli/Makefile:
Added missing path to stampy