Improved cleaning threshold choice
With low coverage data on small graphs, picking a cleaning theshold would often
be unreliable. For instance, poor cleaning thresholds often caused the
`tests/pipeline` test to fail.

I have added two new approaches to pick the cleaning cutoff. The approaches are
tried in turn and we stop when we have sucessfully found a cutoff. Code is a bit
tidier in clean_graph.c with shorter methods.

cleaning_pick_kmer_threshold() doesn't take an FDR parameter anymore. Instead
we use 0.001 and try a different approach if it fails to find a suitable cutoff.
When fitting our error model we now force beta to be >=1 to avoid a poor fit.