Fix breakpoint caller for single samples
breakpoint_caller.c: don't require a fork to find breakpoints. Just look for
a kmer in the ref connected to a kmer not in the ref.
Previously we loaded and merged the reference and sample edges. Splits from the
reference caused unitigs to end. Therefore we only needed to look at the end of
a supernode for a sample splitting from the ref. If we don't infer edges on a
single sample, we can't load the reference edges. This should have no effect on
breakpoint variant calling. However the reference may join in the middle of a
unitig so we need to check for this.

NOTE: Bubble calling will not work with links and SINGLE_SAMPLE=1 due to the use
of the reference as a separate colour.

src/commands/ctx_calls2vcf.c:
* don't name variants in VCF, set ID field to '.'
* breakpoint calls: if two clamps of the same size, take the one closest to
  the breakpoint.

scripts/make-pipeline.pl:
* fix bug that was only passing --no-ref-edges to the breakpoint caller when it
  shouldn't.
* bcftools needs to call norm twice to normalise then remove duplicates
* don't rename variants during VCF filtering

kmer_occur.h:
* rename kmer_run_buf_*() -> korun_buf_*()
* lots of refactoring of public methods

graph_cache.h:
* rename GCacheSnode -> GCacheUnitig
* rename cache_snode_buf_*() -> cache_unitig_buf_*()

src/basic/hash.h: gather hash functions under ctx_hash32(), ctx_hash64()

scripts/R/plot-covg-hist.R: use Rscript for easier calling

db_node.h: added db_nodes_get() to get a node from a contig with orientation