Fix mccortex-server.py to exit cleanly
mccortex-server.py was not exiting cleaning on SIGINT (^C) due to the signal
being passed to the subprocess which was being killed by it. We now catch the
SIGINT instead of passing it to the subprocess, and give bin/mccortex31 a chance
to shutdown cleanly by printing an exit command to it ("q").