Fix multithreading bug in supernodes_iterate()
Fix bug in supernodes_iterate() caused by different locks being used by
different threads to access the same cyclic unitig. When we normalise a cyclic
unitig, we cyclic shift the nodes rather than just flipping it. We were not
taking this into account when picking a key node to choose the correct lock.
Key node is the node with the lowest kmer index in the hashtable.